lib+common: Fix v.h linkage for dynamically linked build
`make lib` builds c-ray in two parts, `lib/libc-ray.so` and `lib/c-ray`.
When I added v.h initially, I just placed the symbols in
`lib/api/c-ray.c`, but doing it that way failed to export the symbols
for the driver portion in `lib/c-ray`. Fix is to put the symbols in a
new translation unit under `src/common`.